platform-sign-pdf/hello_world/requirements.txt
requests
asn1crypto==0.23.0
attrs==19.2.0
cffi==1.12.3
cryptography==2.7
endesive==1.2.0
fonttools==4.0.2
lxml==4.4.1
pdf-annotate==0.11.0
pdfminer.six==20181108
pdfrw==0.4
Pillow==6.2.0
pycparser==2.19
pycryptodome==3.9.0
PyKCS11==1.5.5
pyOpenSSL==19.0.0
six==1.12.0
sortedcontainers==2.1.0
boto3==1.10.19
botocore==1.13.19